move payment component into pages/payments folder
